+++ NEWS.md
@@ -1,3 +1,23 @@
+# mrgsolve 0.11.2
+
+- In `pk2iv`, change scaling volume for `CENT` from `V2` (incorrect) to 
+  `V1` (#831, #832, #833)
+  
+- Fix bug collating multiple `$OMEGA` or `$SIGMA` matrices when parsing a 
+  model (#863)
+  
+- Refactor how debugging information is processed when using the `recover` 
+  argument to `mread()` (#853)
+  
+- Fix typo in documentation for `as_cmat()`; the off-diagonals are assumed
+  to contain correlations (#856)
+
+- Wrote a `as.data.frame.matrix()` function in Rcpp; all simulation results
+  return from the C++ simulation code as a data frame (#857)
+  
+- Fix bug where dynamic capture (via `mread()`) was not allowed for variables
+  declared in `$GLOBAL` (#868)
+
 # mrgsolve 0.11.1
 
 - `mrgsim()` will now periodically check for user interrupt signal so that